+++ tests/acceptance/components/flexberry-objectlistview/folv-date-format-moment-test.js
@@ -1,12 +1,12 @@
 import $ from 'jquery';
 import { get } from '@ember/object';
 import { executeTest } from './execute-folv-test';
-import { loadingList, loadingLocales } from './folv-tests-functions';
+import { loadingLocales, refreshListByFunction } from './folv-tests-functions';
 
 import I18nRuLocale from 'ember-flexberry/locales/ru/translations';
 
 executeTest('date format moment L', (store, assert, app) => {
-  assert.expect(7);
+  assert.expect(5);
   let done = assert.async();
   let path = 'components-acceptance-tests/flexberry-objectlistview/base-operations';
   visit(path);
@@ -15,16 +15,19 @@ executeTest('date format moment L', (store, assert, app) => {
     loadingLocales('ru', app).then(() => {
 
       let olvContainerClass = '.object-list-view-container';
-      let trTableClass = 'table.object-list-view tbody tr';
 
       let $toolBar = $('.ui.secondary.menu')[0];
       let $toolBarButtons = $toolBar.children;
       let $refreshButton = $toolBarButtons[0];
       assert.equal($refreshButton.innerText.trim(), get(I18nRuLocale, 'components.olv-toolbar.refresh-button-text'), 'button refresh exist');
 
-      loadingList($refreshButton, olvContainerClass, trTableClass).then(($list) => {
-        assert.ok($list, 'list loaded');
+      let controller = app.__container__.lookup('controller:' + currentRouteName());
+      let refreshFunction =  function() {
+        let refreshButton = Ember.$('.refresh-button')[0];
+        refreshButton.click();
+      };
 
+      refreshListByFunction(refreshFunction, controller).then(() => {
         let moment = app.__container__.lookup('service:moment');
         let momentValue = get(moment, 'defaultFormat');
 
@@ -61,9 +64,7 @@ executeTest('date format moment L', (store, assert, app) => {
         loadingLocales('en', app).then(() => {
 
           let done1 = assert.async();
-          loadingList($refreshButton, olvContainerClass, trTableClass).then(($list) => {
-            assert.ok($list, 'list loaded');
-
+          refreshListByFunction(refreshFunction, controller).then(() => {
             // Date format most be MM/DD/YYYY:
             let dateFormatEnRe = /(0[1-9]|1[012])\/(0[1-9]|[12][0-9]|3[01])\/(19|20)\d\d/;
             let dataCellStr = $dateCell();