+++ tests/acceptance/components/flexberry-objectlistview/folv-delete-button-test.js
@@ -2,7 +2,7 @@ import { run } from '@ember/runloop';
 import $ from 'jquery';
 import RSVP from 'rsvp';
 import { A } from '@ember/array';
-import { executeTest } from './execute-folv-test';
+import { executeTest, addDataForDestroy } from './execute-folv-test';
 import { loadingList } from './folv-tests-functions';
 
 import generateUniqueId from 'ember-flexberry-data/utils/generate-unique-id';
@@ -32,6 +32,8 @@ executeTest('check delete using button on toolbar', (store, assert, app) => {
     newRecords.forEach(function(item) {
       promises.push(item.save());
     });
+
+    addDataForDestroy(newRecords);
 
     RSVP.Promise.all(promises).then(function(resolvedPromises) {
       assert.ok(resolvedPromises, 'All records saved.');