+++ tests/acceptance/components/flexberry-objectlistview/folv-sorting-clear-test.js
@@ -41,7 +41,7 @@ executeTest('check sorting clear', (store, assert, app) => {
           assert.equal($thead.children[0].children.length, 1, 'no sorting icon in the first column');
 
           // Refresh function.
-          let refreshFunction =  function() {
+          let refreshFunction1 =  function() {
             $thead.click();
           };
 
@@ -58,12 +58,13 @@ executeTest('check sorting clear', (store, assert, app) => {
             checkSortingList(store, projectionName, $olv, 'address asc').then((isTrue) => {
               assert.ok(isTrue, 'sorting applied');
 
-              let $clearButton = $('.clear-sorting-button');
-              $clearButton.click();
-
               let done3 = assert.async();
+              let refreshFunction2 =  function() {
+                let $clearButton = $('.clear-sorting-button');
+                $clearButton.click();
+              };
 
-              window.setTimeout(() => {
+              refreshListByFunction(refreshFunction2, controller).then(() => {
                 let $thead = $('th.dt-head-left', $olv)[0];
                 let $ord = $('.object-list-view-order-icon', $thead);
                 let $divOrd = $('div', $ord);
@@ -72,7 +73,7 @@ executeTest('check sorting clear', (store, assert, app) => {
                 assert.equal($.trim($divOrd.text()), '', 'sorting symbol delete');
 
                 done3();
-              }, 3000);
+              });
               done2();
             });
             done1();