+++ tests/dummy/app/controllers/components-examples/flexberry-lookup/dropdown-mode-example.js
@@ -1,6 +1,4 @@
-import { computed } from '@ember/object';
 import EditFormController from 'ember-flexberry/controllers/edit-form';
-import { StringPredicate } from 'ember-flexberry-data/query/predicate';
 
 export default EditFormController.extend({
 });