+++ tests/dummy/app/locales/ru/translations.js
@@ -739,6 +739,10 @@ $.extend(true, translations, {
               'caption': 'Разметка столбцов',
               'title': ''
             },
+            'return-from-ediform': {
+              'title': 'Возвращение с edit-form на list-form с queryParametr',
+              'return-button': 'Вернутся'
+            },
             'lock-services-editor-view-list': {
               'caption': 'Пример отображение имени пользователя заблокировшего объект',
               'title': ''
@@ -773,6 +777,10 @@ $.extend(true, translations, {
             'title': '',
             'settings-example': {
               'caption': 'Пример работы с настройками',
+              'title': ''
+            },
+            'settings-example-inner': {
+              'caption': 'Пример работы с настройками (toggler в toggler)',
               'title': ''
             },
             'ge-into-toggler-example': {
@@ -1173,6 +1181,11 @@ $.extend(true, translations, {
         'settings-example': {
           'caption': 'Flexberry-toggler. Пример работы с настройками',
           'togglerContent': 'Некоторое  сворачивающееся/разворачивающееся содержание'
+        },
+        'settings-example-inner': {
+          'caption': 'Flexberry-toggler. Пример работы с настройками',
+          'togglerContent': 'Некоторое  сворачивающееся/разворачивающееся содержание',
+          'innerTogglerContent': 'Некоторое  сворачивающееся/разворачивающееся содержание внутреннего тогглера'
         },
         'ge-into-toggler-example': {
           'caption': 'Flexberry-toggler. GroupEdit в toggler example'